MED-169 feat: separating modal of appointment
