+++ apps/medon-fe/src/pages/AppointmentsPageModal/types.ts
@@ -1,0 +1,7 @@
+import { Event } from 'react-big-calendar';
+
+export interface IAppointmentsPageModalProps {
+  isVisible: boolean;
+  hideModal: () => void;
+  selectedEvent: Event | null;
+}
